+++ changelog.md
@@ -4,6 +4,8 @@
   - Update API version 1.2.135
   - Bracket functions now take as an argument a function to consume a pair of
     begin/end actions
+  - Change `Either Word32 vec` to `vec` where the non-empty length is
+    constrained to be the length of another known vector
   - Move `Counter` prefix on `PerformanceCounterResult` to be a suffix
 
 ## [3.1.0.0] - 2020-03-10